requirements.txt
flask
google-generativeai
pymongo
requests
gunicorn